Move contains and fillna. Modify _delegate_property logic to support inplace for fillna. Delete _row_wise_operation in GeoSeries
